requirements.txt
blinker==1.6.2
click==8.1.4
colorama==0.4.6
dominate==2.8.0
Flask==2.3.2
Flask-Bootstrap4==4.0.2
flask-bootstrap5==0.1.dev1
Flask-MySQL==1.5.2
Flask-MySQLdb==1.0.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
PyMySQL==1.1.0
PyYAML==6.0
SQLAlchemy==2.0.18
typing_extensions==4.7.1
visitor==0.1.3
Werkzeug==2.3.6
